refactor(chat): split routes.py into routes/ package
Split the monolithic routes.py (~1376 lines) into focused submodules:
- routes/dependencies.py  — FastAPI dependency factories
- routes/completions.py   — /completions endpoint + stream_chat_response
- routes/conversations.py — conversation CRUD endpoints
- routes/attachments.py   — attachment management endpoints
- routes/tool_endpoints.py — /tools endpoint
- routes/__init__.py      — assembles chat_router, re-exports public API via __all__

Updated mock patch targets in tests from routes.* to the correct
submodule paths (routes.completions.*, routes.dependencies.*).
