+++ app/core_plugins/chat/routes/conversations.py
@@ -1,0 +1,185 @@
+import json
+from typing import Any
+from uuid import UUID
+
+from fastapi import APIRouter, Depends, HTTPException, Query, status
+from sqlmodel import col, func, select
+from sqlmodel.ext.asyncio.session import AsyncSession
+
+from app.api.v1.auth import get_current_user
+from app.core.db import get_async_session
+from app.core.logger import get_logger
+from app.core_plugins.chat.models import Message
+from app.core_plugins.chat.routes.dependencies import get_chat_service
+from app.core_plugins.chat.schemas import (
+    ConversationDetailResponse,
+    ConversationListResponse,
+    ConversationResponse,
+    MessageResponse,
+)
+from app.core_plugins.chat.service import ChatService
+from app.models.user import User
+
+logger = get_logger(__name__)
+
+router = APIRouter()
+
+
+def _parse_metadata_list(model_metadata: str | None, key: str) -> list[dict[str, Any]] | None:
+    if not model_metadata:
+        return None
+    try:
+        meta = json.loads(model_metadata)
+        value = meta.get(key)
+        return value if isinstance(value, list) else None
+    except (json.JSONDecodeError, AttributeError) as exc:
+        logger.error("Failed to parse model_metadata for key %r: %s", key, exc)
+        return None
+
+
+@router.get("/conversations", response_model=ConversationListResponse)
+async def list_conversations(
+    limit: int = Query(default=50, ge=1, le=200),
+    offset: int = Query(default=0, ge=0),
+    current_user: User = Depends(get_current_user),
+    session: AsyncSession = Depends(get_async_session),
+    service: ChatService = Depends(get_chat_service),
+) -> ConversationListResponse:
+    conversations, total = await service.list_conversations(
+        session=session,
+        user_id=current_user.id,  # type: ignore
+        limit=limit,
+        offset=offset,
+    )
+
+    conv_ids = [conv.id for conv in conversations]
+
+    count_stmt = (
+        select(Message.conversation_id, func.count(col(Message.id)).label("message_count"))
+        .where(col(Message.conversation_id).in_(conv_ids))
+        .group_by(col(Message.conversation_id))
+    )
+    count_result = await session.exec(count_stmt)
+    message_counts = {row[0]: row[1] for row in count_result.all()}
+
+    conversation_responses = [
+        ConversationResponse(
+            id=conv.uuid,
+            provider=conv.provider,
+            model=conv.model,
+            title=conv.title,
+            total_tokens_used=conv.total_tokens_used,
+            total_cost=conv.total_cost,
+            message_count=message_counts.get(conv.id, 0),  # type: ignore
+            created_at=conv.created_at,
+            updated_at=conv.updated_at,
+        )
+        for conv in conversations
+    ]
+
+    return ConversationListResponse(
+        conversations=conversation_responses,
+        total=total,
+    )
+
+
+@router.get("/conversations/{conversation_id}", response_model=ConversationDetailResponse)
+async def get_conversation(
+    conversation_id: UUID,
+    limit: int = Query(50, ge=1, le=100),
+    offset: int = Query(0, ge=0),
+    current_user: User = Depends(get_current_user),
+    session: AsyncSession = Depends(get_async_session),
+    service: ChatService = Depends(get_chat_service),
+) -> ConversationDetailResponse:
+    conversation = await service.get_conversation_by_uuid(
+        session=session,
+        uuid=conversation_id,
+        user_id=current_user.id,  # type: ignore
+    )
+
+    if not conversation:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Conversation not found",
+        )
+
+    messages = await service.get_conversation_messages(
+        session=session,
+        conversation_id=conversation.id,  # type: ignore
+        limit=limit,
+        offset=offset,
+        exclude_errors=False,
+    )
+
+    message_count = len(messages)
+
+    return ConversationDetailResponse(
+        id=conversation.uuid,
+        provider=conversation.provider,
+        model=conversation.model,
+        title=conversation.title,
+        total_tokens_used=conversation.total_tokens_used,
+        total_cost=conversation.total_cost,
+        message_count=message_count,
+        created_at=conversation.created_at,
+        updated_at=conversation.updated_at,
+        messages=[
+            MessageResponse(
+                id=msg.id,  # type: ignore
+                role=msg.role,
+                content=msg.content,
+                tokens_used=msg.tokens_used,
+                cost=msg.cost,
+                created_at=msg.created_at,
+                message_type=msg.message_type,
+                attachment_name=msg.attachment_name,
+                attachment_size=msg.attachment_size,
+                rag_sections=_parse_metadata_list(msg.model_metadata, "rag_sections"),
+                tool_calls=_parse_metadata_list(msg.model_metadata, "tool_calls"),
+                is_error=msg.is_error,
+            )
+            for msg in messages
+        ],
+    )
+
+
+@router.get(
+    "/conversations/{conversation_id}/last-message",
+    response_model=MessageResponse | None,
+)
+async def get_last_conversation_message(
+    conversation_id: UUID,
+    current_user: User = Depends(get_current_user),
+    session: AsyncSession = Depends(get_async_session),
+    service: ChatService = Depends(get_chat_service),
+) -> MessageResponse | None:
+    conversation = await service.get_conversation_by_uuid(
+        session=session,
+        uuid=conversation_id,
+        user_id=current_user.id,  # type: ignore
+    )
+    if not conversation:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Conversation not found")
+
+    msg = await service.get_last_conversation_message(
+        session=session,
+        conversation_id=conversation.id,  # type: ignore
+    )
+    if msg is None:
+        return None
+
+    return MessageResponse(
+        id=msg.id,  # type: ignore
+        role=msg.role,
+        content=msg.content,
+        tokens_used=msg.tokens_used,
+        cost=msg.cost,
+        created_at=msg.created_at,
+        message_type=msg.message_type,
+        attachment_name=msg.attachment_name,
+        attachment_size=msg.attachment_size,
+        rag_sections=_parse_metadata_list(msg.model_metadata, "rag_sections"),
+        tool_calls=_parse_metadata_list(msg.model_metadata, "tool_calls"),
+        is_error=msg.is_error,
+    )
